requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiogram==3.5.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.2.0
python-dotenv==1.0.1
loguru==0.7.2
pyyaml==6.0.1
httpx==0.27.0

# RAG & Embeddings
langchain==0.2.1
langchain-community==0.2.1
sentence-transformers==2.7.0 # Для локальных эмбеддингов

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
requests==2.31.0 # Для TestClient в FastAPI
